+++ examples/1d_stencil/1d_stencil_8.cpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2014-2022 Hartmut Kaiser
+//  Copyright (c) 2014-2025 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -145,8 +145,8 @@ public:
       , size_(size)
       , min_index_(0)
     {
-        double base_value = double(initial_value * size);
-        for (std::size_t i = 0; i != size; ++i)
+        double base_value = initial_value * double(size);
+        for (std::ptrdiff_t i = 0; i != static_cast<std::ptrdiff_t>(size); ++i)
             data_[i] = base_value + double(i);
     }
 
@@ -629,13 +629,13 @@ stepper_server::space stepper_server::do_work(
         {
             next[0].then([sem, t](partition&&) {
                 // inform semaphore about new lower limit
-                sem->signal(t);
+                sem->signal(static_cast<std::int64_t>(t));
             });
         }
 
         // suspend if the tree has become too deep, the continuation above
         // will resume this thread once the computation has caught up
-        sem->wait(t);
+        sem->wait(static_cast<std::int64_t>(t));
     }
 
     return U_[nt % 2];
