+++ examples/transpose/transpose_smp_block.cpp
@@ -42,7 +42,7 @@ int hpx_main(hpx::program_options::variables_map& vm)
     verbose = vm.count("verbose") ? true : false;
 
     std::uint64_t bytes =
-        static_cast<std::uint64_t>(2.0 * sizeof(double) * order * order);
+        static_cast<std::uint64_t>(2 * sizeof(double) * order * order);
 
     std::uint64_t block_order = order / num_blocks;
     std::uint64_t col_block_size = order * block_order;
@@ -74,7 +74,8 @@ int hpx_main(hpx::program_options::variables_map& vm)
                 double col_val =
                     COL_SHIFT * static_cast<double>(b * block_order + j);
 
-                A[b][i * block_order + j] = col_val + ROW_SHIFT * i;
+                A[b][i * block_order + j] =
+                    col_val + ROW_SHIFT * static_cast<double>(i);
                 B[b][i * block_order + j] = -1.0;
             }
         }
@@ -132,7 +133,8 @@ int hpx_main(hpx::program_options::variables_map& vm)
         avgtime = avgtime /
             static_cast<double>(
                 (std::max) (iterations - 1, static_cast<std::uint64_t>(1)));
-        std::cout << "Rate (MB/s): " << 1.e-6 * bytes / mintime << ", "
+        std::cout << "Rate (MB/s): "
+                  << 1.e-6 * static_cast<double>(bytes) / mintime << ", "
                   << "Avg time (s): " << avgtime << ", "
                   << "Min time (s): " << mintime << ", "
                   << "Max time (s): " << maxtime << "\n";
@@ -229,7 +231,7 @@ double test_results(std::uint64_t order, std::uint64_t block_order,
             double errsq = 0.0;
             for (std::uint64_t i = 0; i < order; ++i)
             {
-                double col_val = COL_SHIFT * i;
+                double col_val = COL_SHIFT * static_cast<double>(i);
                 for (std::uint64_t j = 0; j < block_order; ++j)
                 {
                     double diff = trans[b][i * block_order + j] -
